requirements.txt
streamlit==1.29.0
mysql-connector-python==8.2.0
pandas==2.1.4
beautifulsoup4==4.12.2
